aadhaar_ocr_fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.109.0
h11==0.16.0
idna==3.10
packaging==25.0
pillow==10.2.0
pydantic==2.11.4
pydantic_core==2.33.2
pytesseract==0.3.10
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.1
starlette==0.35.1
tesseract==0.1.3
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.27.0
